requirements.txt
# RGB Matrix Controller Dependencies

# Core RGB Matrix Libraries
rpi-rgb-led-matrix>=1.0.0
adafruit-circuitpython-rgb-display>=3.10.0

# Image Processing
Pillow>=9.0.0
numpy>=1.21.0

# Web Interface
Flask>=2.0.0
Flask-SocketIO>=5.0.0

# Configuration
PyYAML>=6.0
python-dotenv>=0.19.0

# Utilities
requests>=2.28.0
schedule>=1.1.0

# GPIO Control (Raspberry Pi)
RPi.GPIO>=0.7.1
gpiozero>=1.6.0

# Development and Testing
pytest>=7.0.0
black>=22.0.0

# RGB Matrix Emulator for Development
RGBMatrixEmulator>=0.11.0
pygame>=2.1.0
